// frontend/app/services/employeeService/interface.ts
export const enum EmployeeStatus {
    ACTIVE = 'active',
    DELETED = 'deleted',
    FLAGGED = 'flagged',
}

export const enum EmployeeProfile {
    SOFTWARE_ENGINEER = 'Software Engineer',
    QA_ENGINEER = 'QA Engineer',
    PRODUCT_MANAGER = 'Product Manager',
    ENGINEERING_MANAGER = 'Engineering Manager',
}